pyproject.toml
[tool.poetry]
name = "paperdrop"
version = "0.1.0"
description = ""
authors = ["quuavke"]
readme = ""

[tool.poetry.dependencies]
python = "^3.9"
aiofiles = "22.1.0"
aiohttp = "3.9.5"
aioredis = "2.0.1"
aiosignal = "1.3.1"
aiosqlite = "0.17.0"
alembic = "1.9.2"
amqp = "5.1.1"
annotated-types = "0.6.0"
anyio = "3.6.2"
apache-libcloud = "3.8.0"
argon2-cffi = "23.1.0"
argon2-cffi-bindings = "21.2.0"
asgiref = "3.8.1"
asttokens = "2.4.1"
async-generator = "1.10"
async-timeout = "4.0.2"
asyncpg = "0.27.0"
attrs = "22.2.0"
autoflake = "2.0.0"
babel = "2.11.0"
bcrypt = "4.1.2"
billiard = "3.6.4.0"
casbin = "1.36.0"
celery = "5.2.7"
certifi = "2022.12.7"
cffi = "1.15.1"
charset-normalizer = "3.3.2"
click = "8.1.3"
click-didyoumean = "0.3.0"
click-plugins = "1.1.1"
click-repl = "0.2.0"
colorama = "0.4.6"
cryptography = "39.0.0"
dnspython = "2.3.0"
email-validator = "2.1.1"
exceptiongroup = "1.1.0"
executing = "2.0.1"
fastapi = "0.111.0"
fastapi-admin = "1.0.4"
fastapi-cache2 = "0.2.0"
fastapi-cli = "0.0.3"
fastapi-users = "13.0.0"
fastapi-users-db-sqlalchemy = "4.0.4"
fastapi-amis-admin = "0.4.2"
fastapi-user-auth = "0.4.2"
flower = "2.0.1"
frozenlist = "1.4.1"
greenlet = "2.0.1"
h11 = "0.14.0"
httpcore = "0.16.3"
httptools = "0.5.0"
httpx = "0.23.3"
humanize = "4.9.0"
icecream = "2.1.3"
idna = "3.4"
importlib-metadata = "6.0.0"
importlib-resources = "5.10.2"
iniconfig = "2.0.0"
iso8601 = "1.1.0"
itsdangerous = "2.1.2"
jinja2 = "3.1.2"
kombu = "5.2.4"
makefun = "1.15.0"
mako = "1.2.4"
markdown-it-py = "3.0.0"
markupsafe = "2.1.2"
mdurl = "0.1.2"
multidict = "6.0.5"
mypy-extensions = "1.0.0"
orjson = "3.8.5"
outcome = "1.2.0"
packaging = "23.0"
passlib = "1.7.4"
pendulum = "2.1.2"
pluggy = "1.0.0"
prometheus-client = "0.20.0"
prompt-toolkit = "3.0.36"
psycopg2-binary = "2.9.5"
pwdlib = "0.2.0"
pycparser = "2.21"
pydantic = "2.7.1"
pydantic-core = "2.18.2"
pyflakes = "3.0.1"
pygments = "2.18.0"
pyjwt = "2.8.0"
pypika-tortoise = "0.1.6"
pytest = "7.2.1"
pytest-asyncio = "0.20.3"
python-dateutil = "2.8.2"
python-dotenv = "0.21.0"
python-multipart = "0.0.9"
pytz = "2022.7.1"
pytzdata = "2020.1"
pyyaml = "6.0"
redis = "4.4.2"
requests = "2.31.0"
rfc3986 = "1.5.0"
rich = "13.7.1"
shellingham = "1.5.4"
simpleeval = "0.9.13"
six = "1.16.0"
sniffio = "1.3.0"
sortedcontainers = "2.4.0"
sqladmin = "0.16.1"
sqlalchemy = "2.0.30"
sqlalchemy-file = "0.6.0"
sqlalchemy2-stubs = "0.0.2a32"
sqlalchemy-database = "0.1.0"
sqlmodel = "0.0.18"
sqlmodelx = "0.0.4"
starlette = "0.37.2"
time-machine = "2.14.1"
tomli = "2.0.1"
tornado = "6.4.1"
tortoise-orm = "0.19.2"
typer = "0.12.3"
typing-extensions = "4.11.0"
tzdata = "2024.1"
ujson = "5.7.0"
urllib3 = "2.2.1"
uvicorn = "0.20.0"
vine = "5.0.0"
vt-py = "0.18.2"
watchfiles = "0.18.1"
wcwidth = "0.2.6"
websockets = "10.4"
wtforms = "3.1.2"
yarl = "1.9.4"
zipp = "3.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
